refactor(visit-form-service): Split form and visitInstance routes
Both features move route/middleware declarations and response Zod
schemas into their own *.routes.ts files; controllers keep only
request-handling logic via createXController(service) factories.
module.ts files now create the shared DocumentedRouter and register
routes into it. Completes visit-form-service (visitSchedule was
already split in an earlier commit).

No behavior change — verified full-service OpenAPI doc (9 routes
across 3 controllers) byte-identical before/after, lint clean, all 266
existing tests pass, and live smoke checks against the running
service.
