+++ apps/visit-form-service/src/forms/form.module.ts
@@ -1,12 +1,14 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { FormRepository } from './form.repository';
 import { FormService } from './form.service';
-import { createFormRouter } from './form.controller';
+import { registerFormRoutes } from './form.routes';
 
-/** Composition root for the forms feature: wires repository -> service -> router. */
+/** Composition root for the forms feature: wires repository -> service -> routes. */
 export function createFormModule(prisma: PrismaService): DocumentedRouter {
   const repository = new FormRepository(prisma);
   const service = new FormService(repository);
-  return createFormRouter(service);
+  const doc = createDocumentedRouter();
+  registerFormRoutes(doc, service);
+  return doc;
 }
